// src/types/FreelancerSellerFormType.ts
export type FreelancerSellerFormType = {
    addressLine1: string
    addressLine2: string
    // businessName: string
    cityId: string
    comment: string
    // companyRegistrationNumber: string
    countryId: string
    email: string
    firstName: string
    lastName: string
    password: string
    re_password: string
    phone: string
    postcode: string
    // primaryContactPerson: string
    stateId: string
    type: "Freelancer"
    // vat: string
    dob: string
    proofOfIdentity: string
    privacyPolicy: boolean
}

// type: Freelancer
// firstName: dfhfgh
// lastName: sdfghf
// cityId: 61d57561f6d66b40272bee4a
// stateId: 61c42e173dcee46c2a8e46d7
// comment: gjtyjty
// countryId: 61c2fb83dc7c0d455ba5e68d
// addressLine1: ghf
// addressLine2: fgh
// email: z@gmail.com
// postcode: 1234567
// phone: 1236547895
// proofOfIdentity: (binary)
// privacyPolicy: true
// dob: 2022-7-8